@@ -18,7 +18,7 @@ Gem::Specification.new do |s|
 
   s.required_ruby_version = '>= 2.1'
 
-  s.add_dependency "rails", "~> 4.1"
+  s.add_dependency "rails", "~> 5.0"
   s.add_dependency "kaminari"
   s.add_dependency "nokogiri"
   s.add_dependency "fastimage"
